修复当热度值不是整数时 计算相关度错误的bug 原因是float 强制转int 报错

+++ action/related_sort_action.py
@@ -57,7 +57,7 @@ class RelatedSortAction(object):
 
         try:
             if hot_id:
-                status, weight = RelatedSortAction._related_sort_service.deal_hot(hot_id, int(hot_value), clues_ids, zero_ids, int(article_count), int(vip_count), int(negative_emotion_count))
+                status, weight = RelatedSortAction._related_sort_service.deal_hot(hot_id, float(hot_value), clues_ids, zero_ids, int(article_count), int(vip_count), int(negative_emotion_count))
 
             elif article_id:
                 status, weight = RelatedSortAction._related_sort_service.deal_article(article_id, clues_ids, zero_ids, int(may_invalid), int(vip_count), int(negative_emotion_count))
